ok înainte nu lua curba, cred... v2 (i-am mai dat cleanip un pic

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/AdvancedTeleOp.java
@@ -70,7 +70,7 @@ public class AdvancedTeleOp extends OpMode {
 
     private static class MovementOutput {
         boolean isValid;
-        double turnAngle;
+        int turnAngle;
         double leftMotor;
         double rightMotor;
     }
